requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.0.7
django-environ==0.11.2
pillow==10.4.0
pypng==0.20220715.0
qrcode==7.4.2
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
